Use kramed instead of marky-markdown
This removes the need to use oniguruma, which is a painful binary dep
that takes quite some time to install, and brings the markdown parsing a
little bit closer to GitHub flavoring.

A bit more work is needed to add CSS for the sytanx highlighting
classes, and another highlighter function might be worth trying out.